W-12967331: Fix Split packages between mule-extensions-api and mule-module-extensions-xml-support
